App.js was fixed (mainPage) / adaptibility was fixed / code fixed (props/prototype)
