--- src/Components/Information/Information.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import "./Information.scss";
-
-class Information extends React.Component {
-    render() {
-        return (
-            <div className="Information">
-                <div className="Information__text">
-                    <h2>{this.props.props.article}</h2>
-                    <p>{this.props.props.text}</p>
-                    <div className="Information__btn">{this.props.props.btnText}</div>
-                </div>
-                <div className="Information__img" style={{backgroundImage: this.props.props.url}}></div>
-                {/*move backgroundImage into scss */}
-            </div>
-        );
-    }
-}
-
-export default Information;
