+++ src/Components/mainPage/Banner/Banner.jsx
