+++ src/Components/mainPage/Banner/Banner.scss
