+++ src/Components/mainPage/GroupDirection/GroupSlider/GroupSlider.jsx
