+++ src/Components/mainPage/Information/Information.scss
@@ -1,4 +1,3 @@
-
 .Information {
   display: flex;
   margin: 0 auto;
@@ -9,6 +8,7 @@
   justify-content: space-between;
   align-items: center;
 
+  font-family: Segoe UI, Segoe;
   .Information__text {
     width: 51%;
 
@@ -19,7 +19,6 @@
       font-weight: 700;
       font-style: normal;
       font-size: 32px;
-      font-family: Open Sans;
       line-height: 40px;
     }
 
@@ -30,7 +29,6 @@
       font-weight: 400;
       font-style: normal;
       font-size: 18px;
-      font-family: Open Sans;
       line-height: 28px;
     }
 
@@ -45,7 +43,7 @@
       font-weight: 700;
       font-style: normal;
       font-size: 16px;
-      font-family: Open Sans;
+
       line-height: 24px;
       line-height: 40px;
     }
@@ -53,6 +51,7 @@
 
   .Information__img {
     width: 512px;
+    max-width: 98%;
     height: 378px;
     border-radius: 16px;
     background-position: center;
@@ -70,7 +69,7 @@
     .Information__text {
       order: 1;
       width: 650px;
-      max-width: 95%;
+      max-width: 100%;
       margin: 0 auto;
       text-align: center;
 
@@ -80,14 +79,27 @@
         max-width: inherit;
         margin: 0 auto;
       }
+      h2{
+        text-align:center;
+      }
     }
 
     .Information__img {
-      width: 512px;
-      height: 378px;
-      border-radius: 16px;
-      order: 0;
       text-align: center;
+    }
+  }
+}
+@media (max-width: 600px) {
+  .Information {
+    margin-top: 80px;
+  }
+}
+
+@media (max-width: 375px) {
+  .Information {
+    .Information__img {
+      height:270px;
+      margin-bottom: 108px;
     }
   }
 }
