+++ src/Components/mainPage/Slider/Slider.jsx
