+++ src/Components/mainPage/Slider/SliderContent/SliderContent.jsx
