+++ src/Components/mainPage/Slider/SliderContent/SliderContent.scss
@@ -17,15 +17,16 @@
   }
 
   .content {
-    height: 330px;
+
     color: #fff;
     line-height: 160px;
     text-align: center;
-
     .content__wrapper {
-      height: 100%;
+      height: 330px;
+      border-radius:12px;
       width: 90%;
       margin: 0 auto;
+      background:linear-gradient( rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5) );
       background-position: center 30%;
       background-size: cover;
       background-repeat: no-repeat;
@@ -41,7 +42,7 @@
       }
 
       h2 {
-        font-size: calc(12px + 18 * ((100vw - 320px) / 1600));
+        font-size: 40px;
         font-style: normal;
         font-weight: 500;
         line-height: 40px;
@@ -55,9 +56,10 @@
         width: 149px;
         height: 40px;
         margin: 0 auto;
+        margin-top: auto;
         border: 0px;
         border-radius: 6px;
-        bottom: 0px;
+        margin-bottom: 12px ;
 
         p {
           font-size: 16px;
@@ -73,8 +75,7 @@
   }
 
   .slick-dots {
-    bottom: -55px;
-
+    bottom:-10px;
     li {
       width: 12px;
       height: 12px;
@@ -98,30 +99,61 @@
 }
 
 @media (min-width:0px) and (max-width:880px) {
-  .slick-slider {
-    width: 99%;
-    margin: 0 auto;
-
-    .slick-list {
-      height: 85%;
-      width: 100%;
+  #carousel{
+    .slick-slider {
+      width: 99%;
+      margin: 0 auto;
+  
+      .slick-list {
+        height: 85%;
+        width: 100%;
+      }
+  
+      .slick-prev {
+        display: none !important;
+      }
+  
+      .slick-next {
+        display: none !important;
+      }
     }
-
-    .slick-prev {
-      display: none !important;
-    }
-
-    .slick-next {
-      display: none !important;
+  
+    .content {
+      .content__wrapper {
+        height: 100%;
+        width: 100%;
+      }
     }
   }
-
-  .content {
-    .content__wrapper {
-      height: 100%;
-      width: 100%;
-      border-radius: 12px;
-    }
-  }
-
 }
+@media (min-width: 700px) and (max-width: 1120px){
+  #carousel{
+    .content{
+      .content__wrapper{
+        h2{
+          font-size:24px;
+        }
+      }
+    }
+  }
+}
+@media (max-width: 700px){
+  #carousel{
+    .content{
+      .content__wrapper{
+        h2{
+          font-size:24px;
+        }
+      }
+    }
+  }
+}
+
+@media (min-width: 320px) and (max-width: 568px){
+    #carousel{
+      .content{
+        height:580px;
+      }
+    }
+}
+
