+++ src/Components/mainPage/Slider/slider.scss
